Bug fix: Reading the sources of the target cell rather than the source cell Bug fix: incorrectly storing double as int

+++ sonata/hdf5_lib.hpp
@@ -103,7 +103,7 @@ public:
             throw arb::sonata_dataset_exception(name_, (unsigned)i);
         }
 
-        int r = out[0];
+        double r = out[0];
         delete [] out;
 
         return r;
